C++ Tanks vs Zombies | 04 | Live Training | Unreal Engine https://www.youtube.com/watch?v=vi3ZRD5VcgI&index=4&list=PLZlv_N0_O1gaz3ydgU5wt6c_JtJzwXUKW
Making Zombies
00:14:47 https://youtu.be/vi3ZRD5VcgI?list=PLZlv_N0_O1gaz3ydgU5wt6c_JtJzwXUKW&t=887
to
00:31:40 https://youtu.be/vi3ZRD5VcgI?list=PLZlv_N0_O1gaz3ydgU5wt6c_JtJzwXUKW&t=1902

CODE COMPILES AGAIN!
I have adapted the Zombie.h/.cpp code from the tutorial video to the BatteryCollector code, by replacing the references to the Tank type, that does not exist in this demo, against our BatteryCollector Character type.

I will try use the the Mannequin rig as Zombie in the future revisions.